use uint16_t for data, not doing four MCPs right now

+++ src/buttons.cpp
@@ -44,7 +44,7 @@ int buttons_init(uint interrupt_pin, i2c_inst* i2c, uint8_t addr)
     return 0;
 }
 
-bool buttons_task(uint32_t* inputs)
+bool buttons_task(uint16_t* inputs)
 {
     if(interrupt_mcp) {
         interrupt_mcp = false;
