+++ src/buttons.h
@@ -18,6 +18,6 @@ int buttons_init(uint interrupt_pin, i2c_inst* i2c, uint8_t addr);
  * inputs[out] output var for data from buttons, only updated when changed
  * @return if button data is changed
  */
-bool buttons_task(uint32_t* inputs);
+bool buttons_task(uint16_t* inputs);
 
 #endif // BUTTONS_H
